+++ src/app/(root)/(routes)/mypage/suggestions/components/my-suggestion-list-content/index.tsx
@@ -1,0 +1,3 @@
+import MySuggestionListContent from './MySuggestionListContent'
+
+export default MySuggestionListContent
